requirements.txt
pyrogram>=2.0.106
tgcrypto>=1.2.5
python-telegram-bot>=20.7


yandexcloud>=0.272.0
langchain-community>=0.0.29


faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.10.0
sentence-transformers>=2.2.2